ED01000 サインイン、OA10000 メニュー作成

+++ backboneAuth-oa1-service/src/main/java/net/jagunma/backbone/auth/authmanager/application/usecase/signInCommand/SignInRequest.java
@@ -28,5 +28,5 @@ public interface SignInRequest {
      *
      * @return モード
      */
-    Integer getMode();
+    Short getMode();
 }
